feat: split BitFi adapter into separate bfBTC and bfUSD trackers
